Update README.md
OBJECTIVE
The goal is to develop a method that uses genetic data for disease classification
from samples in the datasets which represent patients , to learn the best model 
from training data and use it to predict the label (class) for each sample in test data
where for each patient 7070 genes expressions (values) are measured in order to 
classify the patient’s disease into one of the following cases: EPD, JPA, MED, MGL, 
RHB.
INTRODUCTION
The DNA microarray technology captures gene expressions of thousands of genes 
simultaneously which results in enormous high dimension data with redundant and 
irrelevant genes which makes the analysis challenging. Therefore gene selection 
techniques like Machine Learning, Data Mining algorithms such as decision trees, 
support vector machines, multilayer perceptron, Bayes classifiers, K-Nearest 
Neighbors Ensemble classifier techniques, and so on are used can be used for 
accuracy in prediction of diseases. 
TECHNIQUES USED FOR PREDICTING DISEASE
 Problem Investigation: 
It’s very crucial to understand the objective of problem and understanding of 
final result in order to achieve goal.
 Data Cleaning :
Enormous amount of data is captured which consists of various irrelevant gene 
data as a large number of statistical tests for finding disease classes results in 
the occurrence of many false discoveries among genes called differentially 
expressed. This problem can further manifest itself in the irreproducibility of 
results of different studies, Therefore cleaning the data and preparing data in 
order to better expose the structure of the prediction problem is most 
important step for analysis.
 Analyzing Data: 
For diagnosing the data we use descriptive statistics( this technique summarize 
characteristics of data set) and visualization for the getting more accurate and 
better understanding of the data.
 Determining Algorithm:
Testing different approaches of algorithms on selected data items and selecting 
the best few to examine further in order to get the most out of well-performing 
algorithms on the data.